requirements.txt
allure-pytest~=2.13.2
pytest~=7.2.0
playwright~=1.40.0
pytest-playwright~=0.4.3
colorlog~=6.8.0
PyYAML~=6.0.1